requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
google-generativeai==0.3.2
